Only run os-net-config if file has content
Instack undercloud allows the user to specify their own
os-net-config overrides file. This file currently cannot be empty
because it causes an exception in this puppet manifest.

This patch skips running os-net-config if the config file has no
data and allows the user a means to disable os-net-config from running
if they choose to do so.

Related bug: #1764507

Change-Id: I69eb812af2d653cb3eb29694ae29b57ebb2de53d