Fixes HA Proxy backend check for ODL
This patch removed listen_options for ODL:
https://review.openstack.org/#/c/562036/

Which introduced a regression where default options were then applied
for ODL, including httpchk.  This does not work with ODL because ODL
will not respond to an HTTP GET without specific paths used.  This patch
adds the correct path that may be used to issue HTTP backend check.

Closes-Bug: 1768037

Change-Id: I60bdfc436044851ac02449c262d382b07b888f79
Signed-off-by: Tim Rozet <trozet@redhat.com>