Neutron wrappers: lookup for THT parameter
For Neutron DHCP & L3 agents, lookup for DockerAdditionalSockets
parameter in THT.
It basically replaces the default hiera that was taken from the docker
profile but this one will be deprecated in this release and removed
soon.

Also, we change the default to be '/var/lib/openstack/docker.sock'
so it looks for the right socket in default deployments.

Change-Id: I39c84c462793bfe819f876583782acbe0619a89d