Install python-neutron-lbaas before starting neutron
When installing onto pre-provisioned servers ("split-stack"), if the
python-neutron-lbaas package is not pre-installed, the deploy would
fail when Neutron LBaaS is enabled because tripleo would start Neutron
before installing the required package.

This commit moves neutron::services::lbaas into an earlier step to
ensure the packages are installed at an appropriate time.

Closes-bug: 1777023
Change-Id: I4aebcb66363918e692703d787d823ea0f39f0388