Discover the gid of the docker group for mistral
The gid of the docker group is discovered by stating the gid of the
file /var/run/docker.sock. The gid of the docker group is not fixed,
so it must match the gid assigned when docker is installed on the
host. If no docker.sock exists, the custom fact will return nil.

This change also makes sure the user and group ensure_resource calls are
tagged so that docker-puppet.py can include those tags explicitly.

Blueprint: container-prepare-workflow

Change-Id: I94fa558480e408f76e69d292b1d84849ddf9a2a2