Fix default value for Gnocchi auth mode
The correct values or auth modes for Gnocchi are 'basic' and 'keystone'.
This patch fixes the 'simple' usage to 'basic'. Note that without this rename
the deployment works because when 'simple' is used the parameter is not used
in config file, so value 'basic' is used by python-collectd-gnocchi, because
it is implicit default.

Change-Id: I05632137ed12c59a41a5219189c431983935d461