Conditionalize rndc changes
The changes to allow configuration of more rndc options had a
cross-repo dependency on the tripleo-heat-templates changes that
made use of the new options.  Unfortunately, the puppet changes
landed first and are now broken because necessary inputs are not
being provided.  This change conditionalizes the new configurations
so that if the parameters aren't passed we won't attempt to do
invalid configuration.

Closes-Bug: #1788907
Change-Id: Ia82f1c6be0d5504498e77a90268cad8abecdeae2
(cherry picked from commit 3add939d4c723dec9cc1c2538ea28271d5769041)