Enable configuration of rndc controls
For ha deployments we need to be able to allow rndc commands from
other controllers.  We don't want to allow rndc commands from
anywhere though, so we need to be able to configure it to only
listen on a specific network (which will be internal_api by default).
This can be done via the controls parameter.

However, because the value is a hash and the top-level key for the
hash is a dynamic value, it is problematic to represent the data
structure in a Heat template.  For our purposes, it is sufficient
to break apart the hash into its component parts and build the hash
in our Puppet manifest.

Change-Id: I2c97ff30c4c59ef32b6b918ba52973451e1b6241
Depends-On: https://review.openstack.org/581898