Remove setting root user for ovn dbs pacemaker bundle
This patch removes the option "--user=root" from the docker options.
Right now the kolla image - ovn-northd which is used for ovn dbs pacemaker
bundle doesn't set the user 'openvswitch'. So even after this patch
is merged, the ovn bundle will be started as root user.

This patch is 1st step towards running ovn dbs pacemaker as a non root user.

It requires changes in
 - in ovn-northd kolla image and
 - in the ovn-dbs.yaml pacemaker service file in T-H-T to give the proper
   permissions to the log and data folders
to run the ovn pacemaker bundle as a non root user which will be handled
in follow up patches.

Related-bug: #1781534
Change-Id: I32f37522b012967f12c459cdf4cf14c43c079231