Make sure that stonith state is enforced before attempting a scaleup
While we start supporting the (1->3) controller scaleup use-case,
we want to make sure that any stonith configuration happens before
we do the scaleup.
Since Stonith gets disabled at step 2 of a deploy and reenabled at step 5
(in the case stonith is indeed configured). With this we make sure that
stonith is off before doing the scaleup. While stonith has no reason
to kick in during a scaleup, this is just semantically more correct
and also just a safer default.

Depends-On: Iac31035da98bd68a5481d97ee3765a99563db49f
Change-Id: I2f8a557e9fc912df081f65b216a4245298876854