[Ocata,Pike,Queens-Only] Fix Cinder's Netapp backend
The cinder::backend::netapp resource has logic to determine whether
deprecated parameters are in use. THT supplies empty strings for default
values, which worked fine with Puppet 3.x, where empty strings are treated
as boolean False. When TripleO upgraded to Puppet 4.x in Ocata, the empty
strings are now treated as boolean True. This results in an invalid
configuration, which causes the Netapp backend to fail.

This is not an issue in Rocky, where the deprecated parameters are
being removed.

Closes-Bug: #1782376
Change-Id: I38d175361e7ad3fe38c27bbdc6d29dc6587635dc
(cherry picked from commit 3eea8289f0ba91e73631e9698ae8cefed29e2185)