Correct vrrp script for haproxy status
Currently, on a containerized undercloud, we get error messages from
keepalived container:
/usr/bin/systemctl status haproxy.service exited with status 1

Even if the undercloud is currently not in HA, it's still a good thing
to get clean logs.

Note: the command is launched by the keepalived container, hence we
cannot use pgrep nor anything else. We also need the /var/lib/haproxy
mount in the keepalived container, shared from the haproxy container on
the host.

Change-Id: Ib5050a3a6af262e26b5523974d1d4a80c4e92834