Force cinder properties to be set on ly on nodes with pcmk on it
When using the BlockStorage role there is one cinder volume per node and
those are not managed by pcmk. So when we force the property for all
nodes that have cinder_volume we will fail because pcmk is not running
on those nodes. Let's not set properties for nodes that are not running
pacemaker.

While we're at it let's remove *_nodes_count which are not used anyway.

Closes-Bug: #1786412

Change-Id: I42e9f3244bad60b5df2dfa940f132f72c606620e
(cherry picked from commit 81d4dfa7e0c5d4fcc1cd9fae19b860184c6e351f)