SSL support for haproxy -> novnc proxy connection
With tls-everywhere enabled the connection from haproxy to the nova novnc
proxy was not encrypted. Now we request a certificate and configue haproxy
and the novnc proxy to encrypt this remaining part in a vnc connection to
be encrypted as well.

Change-Id: I4667706633205c240f2efb51663e6efbce5e344e
Related-bug: #1785700
Depends-On: Ice51fe175bdc1cb14fa49cf53d1f38e9728bbb60
(cherry picked from commit 10be0b67c77df7bd510cc8f0b7f4562e24b68cf3)