Fix postrotate to notify holders of rotated logs
Lsof +L1 locates unlinked and open files and does not work for
logrotate, neither with copyteuncate not w/o that option.

Instead, find *.X (X - number) files held and notify the processes
owning those to make an apropriate actions and reopen new log files to
stop writing to the rotated files.

The actions to be taken by such processes are:

* For httpd processes, use USR1 to gracefully reload
* For neutron-server, restart the container as it cannot process
  HUP signal well (LP bug #1276694, LP bug #1780139).
* For nova-compute, restart the container as it cannot process
  HUP signal well (LP bug #1276694, LP bug #1715374).
* For other processes, use HUP to reload

This also fixes the filter to match logfiles ending with *err,
like rabbitmq startup errors log.

Closes-Bug: #1780139
Closes-Bug: #1785659
Closes-Bug: #1715374

Change-Id: I5110426aa26e5fce7ebb4d80d8a2082cbf80519c
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>