Fix fluentd multifile problem.
There is a problem on the current implementation.
Currently the new_source variable is an Array of hashes,
not a hash of hashes. The current impletation is wrong,
as it can be seen for the following error at the log gate:
http://logs.openstack.org/09/563009/7/check/tripleo-ci-centos-7-containers-multinode/aa4faf9/logs/undercloud/home/zuul/overcloud_deploy.log.txt.gz#_2018-09-03_15_17_39
[search for fluentd_service.pp] as there are too many lines at the same timestamp.

Error: Evaluation Error: Error while evaluating a Resource
Statement, Evaluation Error: Error while evaluating a Function
Call, comparison of Hash with Hash failed at
/etc/puppet/modules/fluentd/manifests/config.pp:6:16  at
/etc/puppet/modules/tripleo/manifests/profile/base/logging/fluentd/fluentd_service.pp:47
on node centos-7-rax-iad-0001722758.localdomain

Change-Id: Ifd5fbf6509addf4a564ff83c4551525c9a139ff4