Fix ceph-nfs duplicate property
In change I78c16c9fb28211a7a8ec2187ae76b9a072b76ea2 we eliminated
startup races with ceph nfs, but we need to make sure we do
not create duplicate properies or puppet will fail.

While we're at it we also point the bootstrapnode at
ceph_nfs_short_bootstrap_node_name instead of which is more correct.
It is functionally the same if ceph-nfs and manila-share run on the
same nodes, which is likely the case in most cases, but best
make it correct while we're at it.

Closes-Bug: #1798020

Change-Id: I5a41f95b2d91a25fe617afb9ada0b5201d2b7ef7