Update functions to fix unit tests
With the release of 5.5.7, some of the legacy function definitions no
longer pass in unit tests. This change updates the functions that are
failing in the tests to the modern style (4.x) for function
declarations.  Additionally we're removing teh lookup_hiera_hash
function which is failing but not actually consumed by our code base.

There will be a followup patch to migrate the rest of the parser
functions to the new format, but this patch should unblock the gates.

NOTE: git thinks some of these files have been added/deleted rather than
git move due to the large amount of changes between the two versions of
the file.

Change-Id: Ie7316fd422bd4a5eb91f94016977e5d8d76c27bc
Closes-Bug: #1799786