Enable disablereuse option in Zaqar proxy backend
Modify the Zaqar mod_proxy backend to use use the parameter
disablereuse=On, disabling the ability to reuse connections to the
proxy backend.  This is necessary to prevent reuse of connections by
mod_proxy to the backend Zaqar WebSocket server when the Zaqar
WebSocket server closes the socket on its own, that mod_proxy
occasionally expects to be open.  This creates a new connection with
every request.

Change-Id: Id5f22d8682805566eadd8205a5aa88524ce94b80
Closes-Bug: #1803422