Include octavia::controller
Commits https://review.openstack.org/#/c/612395/ and
https://review.openstack.org/#/c/619013/ deprecated some
parameters in class octavia::worker and moved them to the new class
octavia::controller. The latter commit also removed the include of the
controller class so because of that we have to include it in the Octavia
services manifest files in puppet-tripleo.

Partial-Bug: #1805345

Change-Id: I0256e125de367615c21343683ed5311521d8337c
(cherry picked from commit 43ffaed334e0cdf34329972085ffadd7a2fa1ba0)