Fix new mariadb ipv6 issue
It seems that pre 10.3 mariadb did not except square brackets when
setting the wsrep gcomm address, whereas 10.3 requires them.
Also 10.3 seems to have isssues binding to a specified ipv6 address.
For the time being, until we investigate more in detail, let's just
not bind an ip address for the the gcomm address when using ipv6.

This will unblock all promotions while we get to the bottom of the
issue.

Change-Id: I0b49019065c71edc4497c777f33af7926e8b1238
Closes-Bug: #1808536