Prevent systemd unit files to be created to restart services
This patch removes the code that create/modify Systemd unit files so the
services restart. We did it after Newton when we switched the services
to be managed under Systemd to automatically restart.

With containers, we don't want / need that anymore.
In Pike, Queens and Rocky, containers are restarted by the Docker
Engine.
In Stein, containers are restarted by Systemd (which restart podman
containers) but the unit files are now managed by Paunch.

To avoid weird behaviors, let's clean this up.

Change-Id: I82dca36b3c9ebf4243d270ac91ffbf6ad56d6ca6