Filter apache_certificates_specs if hostname is empty
Not all roles have all networks, when creating certificates
we need to filter out certificate specs where the hiera
lookup returns 'nil'.

Previously we created noop resources with fallback to the
ctlplane network for all networks defined in network data.
So the hiera lookup returned a hostname for all networks
no matter if the roles data included that network. This is
no longer the case, so we need to filter.

Bonus, we no longer create certificates that are'nt actually
used.

Change-Id: I651919488cb68b0b9878b4e21ab376bfc6e3f0fe
Closes-Bug: #1811207