Loadbalance keystone's admin endpoints towards the internal backends
Instead of trying to loadbalance keystone's admin endpoint frontends
towards the backends of the same network. We instead forward them
towards the IP's where the public/internal endpoints are listening on.

This way we definitely don't use the admin endpoint (a step towards
getting rid of it), and we won't get SSL host failures when trying to
access the one single endpoint.

Change-Id: I577373b5ff110c47bc4dc350ca753b7a1957554f