Expose innodb_lock_wait_timeout via hieradata.
Along the same lines of https://review.opendev.org/#/c/603996/.
This commits expoxes one more knob to the operators, allowing them
to tweak the value of innodb_lock_wait_timeout (default = 50).
From upstream docs:

~~~
Time in seconds that an InnoDB transaction waits for an InnoDB
row lock (not table lock) before giving up with the error
ERROR 1205 (HY000): Lock wait timeout exceeded; try restarting transaction.
When this occurs, the statement (not transaction) is rolled back.
~~~

As for the other parameters it defaults to 'undef' so to keep the
default value of 50.

Change-Id: Ic2e1666ad39762caf9bf68272fdb00ee42658bb8