Let fluentd not create /etc/rsyslog.d/fluentd.conf file
Currently, when fluentd is configured to listen to syslogs, Fluentd
configures itself and also rsyslog creating the file
/etc/rsyslog.d/fluentd.conf. This generates an issue when
is deployed on RHEL BZ #1701726.

To avoid this issue, a fluentd_managed_rsyslog parameter is added so
fluentd can be configured to create the file or not.

Change-Id: I388180dc991926ff30f8bbc556f61447152f8dc9
(cherry picked from commit a87f8be7ef9a4be791cccb0018ebf9b3814b162b)