Include 'octavia' driver on ML2/OVN deployments
By default, both the 'amphora' and 'octavia' provider drivers are
enabled in Octavia API. A previous change in puppet-tripleo included the
'ovn' driver when on Neutron ML2/OVN deployments but in doing so omitted
the 'octavia' driver. This patch ensures the 'octavia' driver remains
enabled as is needed for Octavia Tempest tests.

Change-Id: I5b29a51d7edace30e6357a2af95d1af4596ddfd3