Make sure we pass the proper new pcs 0.10 variables
With I852d5d7aa8578c45f0c7215827cedd4ea72c8d0b we will now
use the new system to create pacemaker remotes when pcs 0.10
is installed on the system (RHEL/Centos >=8). This new
way of doing things requires the remotes to set up pcsd
and so we need to pass the same pcs user and password that
are used by pacemaker itself to the remote resource.

Tested on an IHA OSP15 deployment and got a successful deployment.

Depends-On: I852d5d7aa8578c45f0c7215827cedd4ea72c8d0b
Change-Id: I02a8c1d618b11f68f9272b9044ff81ac39d6a81b