IHA robustness improvements
This will avoid useless fencing events in case of stonith problems.  The
'compute-unfence-trigger' resource is simply a dummy resource that is
only used to trigger unfence events. If for whatever reason this
resource is having issues on stop (pcmk bug, node overloaded, etc.) it
makes zero sense to have pacemaker do a fencing action against the node.
Let's just block and show the operator the status and be less harsh in
general.

Tested this and I correctly get the following:
[root@controller-0 ~]# pcs resource show compute-unfence-trigger-clone
 Clone: compute-unfence-trigger-clone
  Resource: compute-unfence-trigger (class=ocf provider=pacemaker type=Dummy)
   Meta Attrs: requires=unfencing
   Operations: migrate_from interval=0s timeout=20 (compute-unfence-trigger-migrate_from-interval-0s)
               migrate_to interval=0s timeout=20 (compute-unfence-trigger-migrate_to-interval-0s)
               monitor interval=10 timeout=20 (compute-unfence-trigger-monitor-interval-10)
               reload interval=0s timeout=20 (compute-unfence-trigger-reload-interval-0s)
               start interval=0s timeout=20 (compute-unfence-trigger-start-interval-0s)
               stop interval=0s on-fail=block timeout=20 (compute-unfence-trigger-stop-interval-0s)

Closes-Bug: #1831234

Change-Id: Ib4884078d54c25da998495ba3e8d47b5e17010ce