Don't require memcached_node_ips when deploying keystone
The keystone module always expects memcached_node_ips to be defined
in hieradata, the value of which is used to configure mod_auth_openidc
when OpenID Connect is enabled for federation.  In some cases, such as
when using a trimmed down custom Controller role for development purposes,
memcached may not even be deployed.  This will result in memcached_node_ips
not being set, which causes a deployment failure.

This patch defaults memcached_ips to an empty list, which allows a
deployment of keystone to succeed even when memcached is not being
deployed.

Change-Id: If44b6d11f8c41c96bd823c3e38bacdc08034986d