Explicitly disable all MySQL/MariaDB query_cache use
The query_cache in MySQL / MariaDB is generally not recommended,
as it tends to produce more cache misses than hits and adds latency to all
queries.  Upstream MySQL / MariaDB are  moving away from it, MariaDB > 10.1.7
now defaults it to OFF, MySQL has deprecated the entire feature. Ideally
puppet-mysql would default query_cache_type to zero as well. While it is likely
that recent TripleO versions are producing a     MariaDB without a query cache
in any case due to its default of  off, add the values explicitly to avoid
confusion and to  support potential backporting efforts.

References:

https://mariadb.com/kb/en/library/mariadb-memory-allocation/#query-cache
https://mysqlserverteam.com/mysql-8-0-retiring-support-for-the-query-cache/
https://mariadb.com/kb/en/library/mariadb-1017-release-notes/#notable-changes
https://haydenjames.io/mysql-query-cache-size-performance/

Change-Id: Iadee0cbffbb50d20a5e97c8883fd955fa5cb9266