Patch fluentd module for unit tests until code lands
The fluentd tdagent provider is currently wrapped in parser function
rather than just being defined normally. We have a proposed patch to the
main module, but until that time let's pull the fix in for the unit
testing. We don't actually consume this provider in tripleo so it
doesn't affect deployment and only shows up in unit testing.

Change-Id: I73c7e73d3c4f77fcb93bca54157031db0a548b61
Closes-Bug: #1833724